+++ data/exposed-jasypt/src/main/kotlin/io/bluetape4k/exposed/sql/jasypt/JasyptBinaryColumnType.kt
@@ -1,27 +1,27 @@
-package io.bluetape4k.exposed.jasypt
+package io.bluetape4k.exposed.sql.jasypt
 
+import org.jetbrains.exposed.sql.BinaryColumnType
 import org.jetbrains.exposed.sql.ColumnTransformer
 import org.jetbrains.exposed.sql.ColumnWithTransform
-import org.jetbrains.exposed.sql.VarCharColumnType
 
-class JasyptVarCharColumnType(
+class JasyptBinaryColumnType(
     private val encryptor: io.bluetape4k.crypto.encrypt.Encryptor,
-    colLength: Int,
-): ColumnWithTransform<String, String>(VarCharColumnType(colLength), JasyptStringEncryptionTransformer(encryptor))
+    length: Int,
+): ColumnWithTransform<ByteArray, ByteArray>(BinaryColumnType(length), JasyptByteArrayEncryptionTransformer(encryptor))
 
-class JasyptStringEncryptionTransformer(
+class JasyptByteArrayEncryptionTransformer(
     private val encryptor: io.bluetape4k.crypto.encrypt.Encryptor,
-): ColumnTransformer<String, String> {
+): ColumnTransformer<ByteArray, ByteArray> {
     /**
      * Encrypts the given value using the provided [encryptor].
      *
      * @param value The value to encrypt.
      * @return The encrypted value.
      */
-    override fun unwrap(value: String) = encryptor.encrypt(value)
+    override fun unwrap(value: ByteArray) = encryptor.encrypt(value)
 
     /**
      * Decrypts the given value using the provided [encryptor].
      */
-    override fun wrap(value: String) = encryptor.decrypt(value)
+    override fun wrap(value: ByteArray) = encryptor.decrypt(value)
 }
