+++ data/exposed-jasypt/src/main/kotlin/io/bluetape4k/exposed/sql/jasypt/JasyptVarCharColumnType.kt
@@ -1,27 +1,27 @@
-package io.bluetape4k.exposed.jasypt
+package io.bluetape4k.exposed.sql.jasypt
 
-import org.jetbrains.exposed.sql.BinaryColumnType
 import org.jetbrains.exposed.sql.ColumnTransformer
 import org.jetbrains.exposed.sql.ColumnWithTransform
+import org.jetbrains.exposed.sql.VarCharColumnType
 
-class JasyptBinaryColumnType(
+class JasyptVarCharColumnType(
     private val encryptor: io.bluetape4k.crypto.encrypt.Encryptor,
-    length: Int,
-): ColumnWithTransform<ByteArray, ByteArray>(BinaryColumnType(length), JasyptByteArrayEncryptionTransformer(encryptor))
+    colLength: Int,
+): ColumnWithTransform<String, String>(VarCharColumnType(colLength), JasyptStringEncryptionTransformer(encryptor))
 
-class JasyptByteArrayEncryptionTransformer(
+class JasyptStringEncryptionTransformer(
     private val encryptor: io.bluetape4k.crypto.encrypt.Encryptor,
-): ColumnTransformer<ByteArray, ByteArray> {
+): ColumnTransformer<String, String> {
     /**
      * Encrypts the given value using the provided [encryptor].
      *
      * @param value The value to encrypt.
      * @return The encrypted value.
      */
-    override fun unwrap(value: ByteArray) = encryptor.encrypt(value)
+    override fun unwrap(value: String) = encryptor.encrypt(value)
 
     /**
      * Decrypts the given value using the provided [encryptor].
      */
-    override fun wrap(value: ByteArray) = encryptor.decrypt(value)
+    override fun wrap(value: String) = encryptor.decrypt(value)
 }
