+++ data/exposed-jasypt/src/test/kotlin/io/bluetape4k/exposed/sql/jasypt/JasyptColumnTypeDaoTest.kt
@@ -1,4 +1,4 @@
-package io.bluetape4k.exposed.jasypt
+package io.bluetape4k.exposed.sql.jasypt
 
 import io.bluetape4k.crypto.encrypt.Encryptors
 import io.bluetape4k.exposed.dao.idEquals
